+++ java/src/com/ctre/phoenix/Motion/ServoGoStraight.java
@@ -17,15 +17,15 @@ public class ServoGoStraight implements com.ctre.phoenix.ILoopable
 		servoParams = params;
 	}
 	
-	float _y;
-	float _heading;
-	float _previousHeading;
-	float _timeElapsed;
+	double _y;
+	double _heading;
+	double _previousHeading;
+	double _timeElapsed;
 	
 	boolean _isRunning = false;
 	boolean _isDone = false;
 	
-	public ServoGoStraight(ISmartDrivetrain driveTrain, Styles.Smart style, ServoParameters params, float y, float heading)
+	public ServoGoStraight(ISmartDrivetrain driveTrain, Styles.Smart style, ServoParameters params, double y, double heading)
 	{
 		_driveTrain = driveTrain;
 		_selectedStyle = style;
@@ -41,35 +41,35 @@ public class ServoGoStraight implements com.ctre.phoenix.ILoopable
         _selectedStyle = style;
     }
     
-    public void Set(float Y , float heading)
+    public void Set(double Y , double heading)
     {
         _y = Y;
 		_heading = heading;
     }
     
-    public float GetEncoderHeading()
+    public double GetEncoderHeading()
     {
         return _driveTrain.GetEncoderHeading();
     }
 	
 	public void resetIAccum() { servoParams.resetIAccum(); }
 
-    private void GoStraight(float Y, float heading)
+    private void GoStraight(double Y, double heading)
     {
     	if(servoParams.P == 0 && servoParams.I == 0 && servoParams.D == 0)
     		com.ctre.phoenix.CTRLogger.Log(-503, "Servo Go Straight");
     	
         /* Grab encoder heading */
-        float currentHeading = GetEncoderHeading();
+        double currentHeading = GetEncoderHeading();
 
         /* Find angular rate from the encoders */
         _timeElapsed = _myStopWatch.getDuration();
-        float correctionRate = ((currentHeading - _previousHeading) / _timeElapsed);
+        double correctionRate = ((currentHeading - _previousHeading) / _timeElapsed);
         _myStopWatch.start();
 
         /* Heading PID */
-        float headingError = heading - currentHeading;
-        float X = servoParams.PID(headingError, correctionRate);
+        double headingError = heading - currentHeading;
+        double X = servoParams.PID(headingError, correctionRate);
         X = -X;
 
         /* Select control style based on selected style */
